requirements.txt
asgiref==3.5.2
Django==3.2
django-crispy-forms==1.14.0
Pillow==9.2.0
psycopg2==2.9.5
pytz==2022.2.1
sqlparse==0.4.2
